pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1"
requests_html = "^0.10.0"
guessit = "^3.0"
imdbparser = "^1.0"
pymongo = "^3.9"
dnspython = "^1.16"
flask_socketio = "^4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
